scrapper/requirements.txt
selenium
meilisearch
python-dotenv
langdetect
requests
pandas